TKSS-9: Share the same build logics
