+++ build.gradle.kts
@@ -1,0 +1,9 @@
+allprojects {
+
+    repositories {
+        mavenCentral()
+    }
+
+    group = "com.tencent.kona"
+    version = "1.0.2-SNAPSHOT"
+}
